Переименовывает компоненты.
