+++ src/components/catalog-movies-card/catalog-movies-card.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+
+export const CatalogMoviesCard = (props) => {
+  // eslint-disable-next-line react/prop-types
+  const {movieTitle} = props;
+
+  return (
+    <article className="small-movie-card catalog__movies-card">
+      <div className="small-movie-card__image">
+        <img src="img/fantastic-beasts-the-crimes-of-grindelwald.jpg" alt={movieTitle} width="280" height="175" />
+      </div>
+      <h3 className="small-movie-card__title">
+        <a className="small-movie-card__link" href="movie-page.html">{movieTitle}</a>
+      </h3>
+    </article>
+  );
+};
